requirements.txt
moviepy>=2.2.1
librosa>=0.11.0
opencv-python>=4.11.0.86
numpy>=2.2.6
faster_whisper~=1.1.1
faster-whisper~=1.1.1
setuptools~=75.8.0
pymediainfo~=7.0.1
uiautomation~=2.0.25
sympy~=1.14.0
nemo_toolkit[asr]~=2.1.0